+++ crates/report/build.rs
@@ -1,0 +1,8 @@
+fn main() {
+    // Uncomment the following when updating the schema.capnp file
+    // capnpc::CompilerCommand::new()
+    //     .file("schema.capnp")
+    //     .output_path("generated/")
+    //     .run()
+    //     .expect("compiling schema.capnp");
+}